requirements.txt
joblib
lief
matplotlib
numpy
pandas
pefile
Pillow
python_magic
PyYAML
requests
scikit_learn
signify
ssdeep
streamlit
xgboost
yara-python
